requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages

astronomer-cosmos
apache-airflow-providers-snowflake

pandas
requests
jsonschema  # Optional, helps with JSON validation
spotipy
snowflake-snowpark-python
apache-airflow
apache-airflow-providers-snowflake